Extract removeHyperlinks() to its own file, so it can be tested
Pure refactoring - done with the Move facility in WebStorm.
